main commit

+++ Matérias/AED1/Trabalho 1/Exercicio3/lista.h
@@ -8,6 +8,6 @@ float tam_lista(Lista lst);
 void mostra_lista(Lista lst);
 int esvazia_lista(Lista *lst);
 int get_elem_pos(Lista lst, int pos, float *elem);
-int remove_posicao(Lista *lst, int pos);
+float remove_posicao(Lista *lst, int pos);
 int iguais(Lista lst1, Lista lst2);
 Lista intercala(Lista lst1,Lista lst2);
